attached_assets/requirements_1767449997953.txt
discord.py>=2.4.0
aiohttp>=3.9.0
python-dateutil>=2.9.0\n